+++ style.py
@@ -6,7 +6,7 @@ class Style:
     growth_arrow_size: float
     growth_arrow_stroke: str
     growth_arrow_fill: str
-
+    stroke_dasharray: str
     # SVG
     fill: str
     stroke: str
@@ -21,6 +21,7 @@ class Style:
     text_stroke_width: int
     text_fill: str
 
+    weigth: int
     def __init__(self, style=None):
         if style is not None:
             for key, value in style.items():
@@ -70,5 +71,7 @@ class Style:
         default_style.fill = 'black'
         default_style.growth_arrow_fill = 'white'
         default_style.growth_arrow_stroke = 'black'
+        default_style.stroke_dasharray = '3,2'
+        default_style.weigth = 2
 
         return default_style